+++ ChoGGi's Library/Code/Init.lua
@@ -1,3 +1,5 @@
+--~ print("LIBRARY REMOVE TEMP FUNCTION")
+
 -- See LICENSE for terms
 
 local LICENSE = [[