+++ Mods ChoGGi/Fix Long Play Ranches/metadata.lua
@@ -1,0 +1,27 @@
+return PlaceObj("ModDef", {
+	"title", "Fix Long Play Ranches",
+	"id", "ChoGGi_FixLongPlayIssues",
+	"steam_id", "2703896422",
+	"pops_any_uuid", "bda132cb-7c8a-441d-9bc6-9f9ab5c111fd",
+	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.png",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagOther", true,
+	"description", [[
+If you have a ranch in a game with a couple thousand Sols then the animals might stop coming out after a harvest.
+
+Mod option to disable.
+
+https://forum.paradoxplaza.com/forum/threads/surviving-mars-lag-and-unplayable-after-a-certain-sol.1490083/
+
+
+[url=https://steamcommunity.com/sharedfiles/filedetails/?id=2705335465]Fix Sol 2983[/url]
+]],
+})