+++ Mods ChoGGi/Fix Rocket Stuck/metadata.lua
@@ -4,10 +4,10 @@ return PlaceObj("ModDef", {
 	"steam_id", "1567028510",
 	"pops_any_uuid", "cca05f79-5542-40f1-a008-e8fb4e6ecc9a",
 	"lua_revision", 1007000, -- Picard
-	"version", 19,
-	"version_major", 1,
-	"version_minor", 9,
-	"image", "Preview.png",
+	"version", 20,
+	"version_major", 2,
+	"version_minor", 0,
+	"image", "Preview.jpg",
 	"author", "ChoGGi",
 	"code", {
 		"Code/Script.lua",
@@ -28,9 +28,10 @@ Returned expedition rocket giving msg that it's still in orbit.
 Canceled expedition rocket stuck in limbo canceled mode.
 Trade rockets stuck in orbit?
 Previous expedition rocket shows msg it isn't landed when trying to launch.
+Trade rocket ready for takeoff stuck on pad only showing priority button.
 
 
-If this doesn't fix it for you, then I'll need a copy of your saved game.
+[b]If this doesn't fix it for you, then I'll need a copy of your saved game.[/b]
 
 
 Has mod option to disable mod (once the rocket is fixed).