+++ Mods ChoGGi/Pinned Icon Scale/metadata.lua
@@ -4,9 +4,9 @@ return PlaceObj("ModDef", {
 	"steam_id", "2653690138",
 	"pops_any_uuid", "1123bd0b-e4aa-462c-af83-be5d18ee2bef",
 	"lua_revision", 1007000, -- Picard
-	"version", 1,
+	"version", 2,
 	"version_major", 0,
-	"version_minor", 1,
+	"version_minor", 2,
 	"image", "Preview.png",
 	"author", "ChoGGi",
 	"code", {